Change form and listings into svelte component
